backend-python/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
gunicorn==21.2.0
